+++ airflow/contrib/example_dags/example_emr_job_flow_manual_steps.py
@@ -25,10 +25,10 @@ terminating the cluster.
 from datetime import timedelta
 
 from airflow import DAG
-from airflow.contrib.operators.emr_terminate_job_flow_operator import EmrTerminateJobFlowOperator
 from airflow.contrib.sensors.emr_step_sensor import EmrStepSensor
 from airflow.providers.amazon.aws.operators.emr_add_steps import EmrAddStepsOperator
 from airflow.providers.amazon.aws.operators.emr_create_job_flow import EmrCreateJobFlowOperator
+from airflow.providers.amazon.aws.operators.emr_terminate_job_flow import EmrTerminateJobFlowOperator
 from airflow.utils.dates import days_ago
 
 DEFAULT_ARGS = {
