+++ airflow/contrib/operators/emr_terminate_job_flow_operator.py
@@ -16,42 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from airflow.exceptions import AirflowException
-from airflow.models import BaseOperator
-from airflow.providers.amazon.aws.hooks.emr import EmrHook
-from airflow.utils.decorators import apply_defaults
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.operators.emr_terminate_job_flow`."""
 
+import warnings
 
-class EmrTerminateJobFlowOperator(BaseOperator):
-    """
-    Operator to terminate EMR JobFlows.
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.operators.emr_terminate_job_flow import EmrTerminateJobFlowOperator  # noqa
 
-    :param job_flow_id: id of the JobFlow to terminate. (templated)
-    :type job_flow_id: str
-    :param aws_conn_id: aws connection to uses
-    :type aws_conn_id: str
-    """
-    template_fields = ['job_flow_id']
-    template_ext = ()
-    ui_color = '#f9c915'
-
-    @apply_defaults
-    def __init__(
-            self,
-            job_flow_id,
-            aws_conn_id='aws_default',
-            *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.job_flow_id = job_flow_id
-        self.aws_conn_id = aws_conn_id
-
-    def execute(self, context):
-        emr = EmrHook(aws_conn_id=self.aws_conn_id).get_conn()
-
-        self.log.info('Terminating JobFlow %s', self.job_flow_id)
-        response = emr.terminate_job_flows(JobFlowIds=[self.job_flow_id])
-
-        if not response['ResponseMetadata']['HTTPStatusCode'] == 200:
-            raise AirflowException('JobFlow termination failed: %s' % response)
-        else:
-            self.log.info('JobFlow with id %s terminated', self.job_flow_id)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.emr_terminate_job_flow`.",
+    DeprecationWarning, stacklevel=2
+)
